pyproject.toml
[tool.poetry]
name = "pycordio"
version = "0.1.0"
description = ""
authors = ["evanchristians <evanryk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
py-cord = "^2.5.0"
taskipy = "^1.12.2"
python-weather = "^2.0.0"
python-dotenv = "^1.0.1"
pynacl = "^1.5.0"
asyncio = "^3.4.3"
ffmpeg = "^1.4"
yt-dlp = "^2024.3.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
start = "python -m bot"
get_cookies = "python -m yt_dlp --cookies-from-browser=chrome https://www.youtube.com/watch?v=dQw4w9WgXcQ --cookies=cookies.txt"
